Move PCLI only classes to swirlds-cli
