+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/utility/VirtualTerminal.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.util;
+package org.hiero.consensus.pcli.utility;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
